requirements.txt
requests
python-dotenv
pytrends
google-api-python-client
google-auth
google-auth-oauthlib
openai
streamlit
tenacity